added graph_factory
